requirements.txt
aiofiles
aiohttp
aiomysql
apscheduler==3.9.1
arrow==1.1.0
arduino
bidict==0.21.2
blinker==1.4
cachetools
certifi==2022.12.7
cffi
chardet==3.0.4
Click==7.0
cryptography
DBUtils
greenlet==1.1.0
h11==0.9.0
h2==3.1.1
hpack==3.0.0
Hypercorn==0.8.2
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe
multidict==4.5.2
priority==1.3.0
pycparser==2.20
pyflo-lib==2.0.9
pymysql
python-dateutil==2.8.1
python-engineio==3.14.2
Quart==0.10.0
Quart-CORS==0.2.0
requests==2.25.0
six==1.16.0
sortedcontainers==2.1.0
SQLAlchemy==1.4.18
toml==0.10.0
typing-extensions==3.7.4
urllib3==1.26.5
websockets==11.0.3
wsproto==0.15.0
pyzmq
psutil

